==> notebooks/02_journey_analysis.ipynb <==
import pandas as pd

df = pd.read_csv("../data/processed/cleaned_activities.csv")

df.head()


# 1) بناء جميع المسارات لكل الحسابات
paths_df = (
    df.sort_values(["account_id", "activity_order"])
      .groupby("account_id")
      .agg(
          Country=("Country", "first"),
          solution=("solution", "first"),
          path=("types", lambda x: " → ".join(x))
      )
      .reset_index()
)

paths_df.head()


top5_paths = (
    paths_df
    .groupby(["Country", "solution", "path"])
    .size()
    .reset_index(name="count")
    .sort_values(["Country", "solution", "count"], ascending=[True, True, False])
    .groupby(["Country", "solution"])
    .head(5)
)

top5_paths.head(20)


top5_paths.to_csv(
    "../data/processed/top5_paths_by_country_solution.csv",
    index=False
)
