==> notebooks/04_top_actions_aggregates.ipynb <==
import sys
import os

sys.path.append(os.path.abspath(".."))  # يضيف customer_journey إلى PYTHONPATH
from src.system import CustomerJourneySystem


import pandas as pd

df = pd.read_csv("../data/processed/cleaned_activities.csv")
df["types"] = df["types"].fillna("Unknown").astype(str).str.strip()

df = df.sort_values(["account_id", "activity_order"])
df["next_action"] = df.groupby("account_id")["types"].shift(-1)

model_df = df[df["next_action"].notna()].copy()
model_df.head()


top4_by_country = (
    model_df.groupby(["Country", "next_action"])
            .size()
            .reset_index(name="count")
            .sort_values(["Country", "count"], ascending=[True, False])
            .groupby("Country")
            .head(4)
            .reset_index(drop=True)
)


top4_by_country.to_csv("../data/processed/top4_next_actions_by_country.csv", index=False)
top4_by_country.head(20)

top4_by_solution = (
    model_df.groupby(["solution", "next_action"])
            .size()
            .reset_index(name="count")
            .sort_values(["solution", "count"], ascending=[True, False])
            .groupby("solution")
            .head(4)
            .reset_index(drop=True)
)

top4_by_solution.head(20)
top4_by_solution.to_csv("../data/processed/top4_next_actions_by_solution.csv", index=False)
top4_by_solution.head(20)

top4_by_country_solution = (
    model_df.groupby(["Country", "solution", "next_action"])
            .size()
            .reset_index(name="count")
            .sort_values(["Country", "solution", "count"], ascending=[True, True, False])
            .groupby(["Country", "solution"])
            .head(4)
            .reset_index(drop=True)
)
top4_by_country_solution.to_csv("../data/processed/top4_next_actions_by_country_solution.csv", index=False)

top4_by_country_solution.head(30)


from src.system import CustomerJourneySystem

sys = CustomerJourneySystem(processed_dir="../data/processed")

# تجريب أي country/solution (: AT/MRS أو BE/MRS أو BE/Digital)
out = sys.add_account(account_id="0010L00001hVmFnQAK", country="AT", solution="MRS")
out
